Build system overhaul
- Introducing Ant & Poney for dependencies retrieval and CMake configuration
- Introducing CMake install process
- Minimum Cmake required version is now 2.8.7
- Generating a zip package using CPack
- Generating RecastNavigationConfig.cmake
    - http://www.cmake.org/Wiki/CMake/Tutorials/Packaging
    - http://www.cmake.org/Wiki/CMake/Tutorials/Exporting_and_Importing_Targets
    - http://www.cmake.org/Wiki/CMake/Tutorials/How_to_create_a_ProjectConfig.cmake_file
- Renaming DebugUtils to RecastDetourDebugUtils

To enjoy the change, build and install. The installed package (in the specified
CMAKE_INSTALL_PREFIX) contains a 'build/cmake' directory containing
RecastNavigationConfig.cmake.

When using RecastNavigation in a project built with CMake, you just have to do a
find_package(RecastNavigation) and set RecastNavigation_DIR to this directory.
Usable variables and targets are defined in RecastNavigationConfig.cmake.